+++ packages/preview-middleware-client/src/adp/init.ts
@@ -1,5 +1,6 @@
 import log from 'sap/base/Log';
 import type RuntimeAuthoring from 'sap/ui/rta/RuntimeAuthoring';
+
 import init from '../cpe/init';
 import { initDialogs } from './init-dialogs';