+++ packages/preview-middleware-client/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux-private/preview-middleware-client
 
+## 0.4.0
+
+### Minor Changes
+
+-   b023f4cb: Enhancing the preview-middleware with new functionality such as Controller Extension (creating "codeExt" change).
+
+## 0.3.1
+
+### Patch Changes
+
+-   a7eda7c5: Fix and reuse manual mocks in unit test, removing ui5Facade, updating tsconfig
+
 ## 0.3.0
 
 ### Minor Changes